Explicitly specify type for bins

+++ src/conversion_scripts/convert_to_bin.py
@@ -8,9 +8,8 @@ def save_to_bin(path: str):
     for key, value in state_dict.items():
         print(f"Layer: {key}")
         print(f"Shape: {tuple(value.shape)}")
-        # print(f"Matrix: \n{value.cpu().numpy()}\n")
 
-        value.cpu().numpy().tofile(os.path.join("src/weights/processed", f"{key}-{(tuple(value.shape))}.bin"))
+        value.cpu().numpy().tofile(os.path.join("src/weights/processed", f"{key}-{(tuple(value.shape))}.bin"), dtype=np.float32)
 
 if __name__ == "__main__":
     save_to_bin(os.path.join("src/weights/raw", "mlp_model.pth"))
